requirements.txt
pytest
playwright
python-dotenv
requests
jsonschema
pytest-html
pytest-rerunfailures
pytest-playwright